+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- 2026-04-07 Input: Allow text selection in disabled state
 - 2026-04-04 Tabs: Support object-based tabs for stable keys
 - 2026-04-03 MonthRangePicker: Fix popover alignment near right screen edge
 - 2026-03-19 ContextMenu: Add `success` variant